+++ src/components/BotDetails/BotDetailsHeader.tsx
@@ -8,6 +8,7 @@ import {
   Clipboard,
   TouchableOpacity,
   Text,
+  ImageSourcePropType,
 } from 'react-native'
 import {observer, inject} from 'mobx-react/native'
 import {k, width} from '../Global'
@@ -138,15 +139,16 @@ class BotDetailsHeader extends React.Component<Props, State> {
                 </RText>
               </View>
             )}
-            {bot.image && (
-              <LazyImage
-                file={bot.image}
-                source={bot.image.thumbnail}
-                style={styles.botImage as any}
-                resizeMode="contain"
-                placeholder={<View style={[styles.botImage, {backgroundColor: GREY}]} />}
-              />
-            )}
+            {bot.image &&
+              bot.image.thumbnail && (
+                <LazyImage
+                  file={bot.image}
+                  source={bot.image.thumbnail! as ImageSourcePropType}
+                  style={styles.botImage as any}
+                  resizeMode="contain"
+                  placeholder={<View style={[styles.botImage, {backgroundColor: GREY}]} />}
+                />
+              )}
             <Separator style={{marginHorizontal: 5}} />
           </View>
         )}