+++ src/components/people-lists/GeofenceShare.tsx
@@ -4,13 +4,13 @@ import {observer, inject} from 'mobx-react/native'
 import {observable} from 'mobx'
 import {Actions} from 'react-native-router-flux'
 import Screen from '../Screen'
-import SelectableProfileList from '../../store/SelectableProfileList'
 import FriendMultiSelect from './FriendMultiSelect'
 import {colors} from '../../constants'
 import {k} from '../Global'
 import {IWocky, IBot} from 'wocky-client'
 
 import {RText, BottomButton} from '../common'
+import {ISearchStore} from '../../store/SearchStore'
 
 type Props = {
   botId: string
@@ -18,22 +18,19 @@ type Props = {
   notificationStore: any // TODO proper
   store: any // TODO proper type
   analytics?: any
+  searchStore?: ISearchStore
 }
 
-@inject('wocky', 'notificationStore', 'store', 'analytics')
+@inject('wocky', 'notificationStore', 'store', 'analytics', 'searchStore')
 @observer
 class GeofenceShare extends React.Component<Props> {
   static rightButton = (props: Props) => <RightButton {...props} />
-
-  selection = SelectableProfileList.create({})
-
   @observable bot?: IBot
 
   componentDidMount() {
-    this.selection.clear()
-    const {friends, getBot} = this.props.wocky
+    const {getBot, friends} = this.props.wocky!
     this.bot = getBot({id: this.props.botId})
-    this.selection.setList(friends.map(f => ({profile: f})))
+    this.props.searchStore!.localResult.setList(friends.map(f => ({profile: f})))
     if (!this.props.store.sharePresencePrimed) {
       // NOTE: had to add a delay to prevent immediately closing
       setTimeout(() => Actions.sharePresencePrimer(), 2000)
@@ -41,7 +38,7 @@ class GeofenceShare extends React.Component<Props> {
   }
 
   share = async () => {
-    const shareSelect = this.selection.selected.map(sp => sp.id)
+    const shareSelect = this.props.searchStore!.localResult.selected.map(sp => sp.id)
     try {
       // TODO: implement share (no accept needed) later when we restore public bots
       // this.bot!.share(shareSelect, '', 'geofence share')
@@ -62,12 +59,12 @@ class GeofenceShare extends React.Component<Props> {
   }
 
   render() {
-    if (!this.selection) return null
-    const selected = this.selection.selected.length > 0
+    const selection = this.props.searchStore!.localResult
+    const selected = selection.selected.length > 0
     return (
       <Screen>
         <FriendMultiSelect
-          selection={this.selection}
+          selection={selection}
           botTitle={this.bot && this.bot.title ? this.bot.title : ''}
           inviteMessage="To share presence!"
         />