+++ src/store/HomeStore.ts
@@ -1,4 +1,5 @@
 import {types, getType, getParent, applySnapshot} from 'mobx-state-tree'
+import {IObservableArray} from 'mobx'
 import {Bot, IBot, Location} from 'wocky-client'
 
 export const SelectableCard = types
@@ -8,7 +9,7 @@ export const SelectableCard = types
   .views(self => ({
     // returns index within parent list
     get index() {
-      return getParent(self).findIndex(item => {
+      return (getParent(self) as IObservableArray).findIndex(item => {
         return item === self
       })
     },
@@ -16,7 +17,7 @@ export const SelectableCard = types
   .actions(self => ({
     setSelected: (value: boolean) => (self.isSelected = value),
     select: () => {
-      getParent(self, 2).setIndex(self.index)
+      ;(getParent(self, 2) as IHomeStore).setIndex(self.index)
     },
   }))
 
@@ -58,9 +59,9 @@ const HomeStore = types
     creationMode: false,
     homeBotList: types.optional(types.array(Card), [{tutorial: true}, {you: true}]), // pre-populate with 'you', tutorial card
     index: 0,
-    focusedBotLocation: types.maybe(Location),
-    mapCenterLocation: types.maybe(Location),
-    scrolledToBot: types.maybe(types.reference(Bot)),
+    focusedBotLocation: types.maybeNull(Location),
+    mapCenterLocation: types.maybeNull(Location),
+    scrolledToBot: types.maybeNull(types.reference(Bot)),
   })
   .views(self => ({
     // return the list for current mode
@@ -147,10 +148,6 @@ const HomeStore = types
       disableFullScreen: () => {
         self.fullScreenMode = false
       },
-      postProcessSnapshot(snapshot: any) {
-        // No need to persist this store
-        return {}
-      },
       start() {
         // empty
       },
@@ -158,6 +155,10 @@ const HomeStore = types
         self.setFocusedLocation(null) // otherwise focused location will not be changed and reaction will not fire
       },
     }
+  })
+  .postProcessSnapshot((snapshot: any) => {
+    // No need to persist this store
+    return {}
   })
   .actions(self => ({
     selectBot(bot: IBot) {